+++ cvat-core/src/frames.ts
@@ -38,6 +38,7 @@ export class FrameData {
         stopFrame,
         decodeForward,
         deleted,
+        downloadOriginal,
         related_files: relatedFiles,
     }) {
         Object.defineProperties(
@@ -81,6 +82,18 @@ export class FrameData {
                 },
                 deleted: {
                     value: deleted,
+                    writable: false,
+                },
+                /**
+                 * True if frame downloaded is from original data
+                 * @name downloadOriginal
+                 * @type {boolean}
+                 * @memberof module:API.cvat.classes.FrameData
+                 * @readonly
+                 * @instance
+                 */
+                downloadOriginal: {
+                    value: downloadOriginal,
                     writable: false,
                 },
             }),
@@ -158,7 +171,7 @@ FrameData.prototype.data.implementation = async function (onServerRequest) {
             const taskDataCache = frameDataCache[this.jid];
             const activeChunk = taskDataCache.activeChunkRequest;
             activeChunk.request = serverProxy.frames
-                .getData(null, this.jid, activeChunk.chunkNumber)
+                .getData(null, this.jid, activeChunk.chunkNumber, this.downloadOriginal)
                 .then((chunk) => {
                     frameDataCache[this.jid].activeChunkRequest.completed = true;
                     if (!taskDataCache.nextChunkRequest) {
@@ -396,7 +409,7 @@ class FrameBuffer {
         return this._size - Object.keys(this._buffer).length - requestedFrameCount;
     }
 
-    requestOneChunkFrames(chunkIdx) {
+    requestOneChunkFrames(chunkIdx, original) {
         return new Promise((resolve, reject) => {
             this._requestedChunks[chunkIdx] = {
                 ...this._requestedChunks[chunkIdx],
@@ -414,6 +427,7 @@ class FrameBuffer {
                     stopFrame: frameDataCache[this._jobID].stopFrame,
                     decodeForward: false,
                     deleted: requestedFrame in frameDataCache[this._jobID].meta,
+                    downloadOriginal: original,
                 });
 
                 frameData
@@ -442,7 +456,7 @@ class FrameBuffer {
         });
     }
 
-    fillBuffer(startFrame, frameStep = 1, count = null) {
+    fillBuffer(startFrame, frameStep = 1, count = null, original = false) {
         const freeSize = this.getFreeBufferSize();
         const requestedFrameCount = count ? count * frameStep : freeSize * frameStep;
         const stopFrame = Math.min(startFrame + requestedFrameCount, this._stopFrame + 1);
@@ -484,7 +498,7 @@ class FrameBuffer {
         return new Promise(async (resolve, reject) => {
             for (const chunkIdx of Object.keys(this._requestedChunks)) {
                 try {
-                    const chunkFrames = await this.requestOneChunkFrames(chunkIdx);
+                    const chunkFrames = await this.requestOneChunkFrames(chunkIdx, original);
                     if (chunkIdx in this._requestedChunks) {
                         bufferedFrames = new Set([...bufferedFrames, ...chunkFrames]);
 
@@ -508,11 +522,11 @@ class FrameBuffer {
         });
     }
 
-    async makeFillRequest(start, step, count = null) {
+    async makeFillRequest(start, step, count = null, original = false) {
         if (!this._activeFillBufferRequest) {
             this._activeFillBufferRequest = true;
             try {
-                await this.fillBuffer(start, step, count);
+                await this.fillBuffer(start, step, count, original);
                 this._activeFillBufferRequest = false;
             } catch (error) {
                 if (typeof error === 'number' && error in this._requestedChunks) {
@@ -523,7 +537,13 @@ class FrameBuffer {
         }
     }
 
-    async require(frameNumber: number, jobID: number, fillBuffer: boolean, frameStep: number): FrameData {
+    async require(
+        frameNumber: number,
+        jobID: number,
+        fillBuffer: boolean,
+        frameStep: number,
+        downloadOriginal: boolean,
+    ): FrameData {
         for (const frame in this._buffer) {
             if (+frame < frameNumber || +frame >= frameNumber + this._size * frameStep) {
                 delete this._buffer[frame];
@@ -540,6 +560,7 @@ class FrameBuffer {
             stopFrame: frameDataCache[jobID].stopFrame,
             decodeForward: !fillBuffer,
             deleted: frameNumber in frameDataCache[jobID].meta.deleted_frames,
+            downloadOriginal,
         });
 
         if (frameNumber in this._buffer) {
@@ -554,7 +575,7 @@ class FrameBuffer {
             ) {
                 const maxFrame = cachedFrames ? Math.max(...cachedFrames) : frameNumber;
                 if (maxFrame < this._stopFrame) {
-                    this.makeFillRequest(maxFrame + 1, frameStep).catch((e) => {
+                    this.makeFillRequest(maxFrame + 1, frameStep, null, downloadOriginal).catch((e) => {
                         if (e !== 'not needed') {
                             throw e;
                         }
@@ -563,7 +584,7 @@ class FrameBuffer {
             }
         } else if (fillBuffer) {
             this.clear();
-            await this.makeFillRequest(frameNumber, frameStep, fillBuffer ? null : 1);
+            await this.makeFillRequest(frameNumber, frameStep, fillBuffer ? null : 1, downloadOriginal);
             frame = this._buffer[frameNumber];
         } else {
             this.clear();
@@ -652,6 +673,7 @@ export async function getFrame(
     isPlaying: boolean,
     step: number,
     dimension: DimensionType,
+    downloadOriginal: boolean,
 ) {
     if (!(jobID in frameDataCache)) {
         const blockType = chunkType === 'video' ? cvatData.BlockType.MP4VIDEO : cvatData.BlockType.ARCHIVE;
@@ -696,7 +718,7 @@ export async function getFrame(
         frameDataCache[jobID].provider.setRenderSize(frameMeta.width, frameMeta.height);
     }
 
-    return frameDataCache[jobID].frameBuffer.require(frame, jobID, isPlaying, step);
+    return frameDataCache[jobID].frameBuffer.require(frame, jobID, isPlaying, step, downloadOriginal);
 }
 
 export async function getDeletedFrames(instanceType, id) {