+++ cvat-core/src/server-proxy.ts
@@ -1462,7 +1462,7 @@ async function getImageContext(jid, frame) {
     return response.data;
 }
 
-async function getData(tid, jid, chunk) {
+async function getData(tid, jid, chunk, original) {
     const { backendAPI } = config;
 
     const url = jid === null ? `tasks/${tid}/data` : `jobs/${jid}/data`;
@@ -1472,7 +1472,7 @@ async function getData(tid, jid, chunk) {
         response = await workerAxios.get(`${backendAPI}/${url}`, {
             params: {
                 ...enableOrganization(),
-                quality: 'compressed',
+                quality: original ? 'original' : 'compressed',
                 type: 'chunk',
                 number: chunk,
             },