+++ cvat-ui/src/components/header/settings-modal/player-settings.tsx
@@ -26,6 +26,7 @@ interface Props {
     rotateAll: boolean;
     smoothImage: boolean;
     showDeletedFrames: boolean;
+    showOriginalImage: boolean;
     canvasBackgroundColor: string;
     onChangeFrameStep(step: number): void;
     onChangeFrameSpeed(speed: FrameSpeed): void;
@@ -34,6 +35,7 @@ interface Props {
     onChangeCanvasBackgroundColor(color: string): void;
     onSwitchSmoothImage(enabled: boolean): void;
     onSwitchShowingDeletedFrames(enabled: boolean): void;
+    onSwitchShowOriginalImage(enabled: boolean): void;
 }
 
 export default function PlayerSettingsComponent(props: Props): JSX.Element {
@@ -44,6 +46,7 @@ export default function PlayerSettingsComponent(props: Props): JSX.Element {
         rotateAll,
         smoothImage,
         showDeletedFrames,
+        showOriginalImage,
         canvasBackgroundColor,
         onChangeFrameStep,
         onChangeFrameSpeed,
@@ -52,6 +55,7 @@ export default function PlayerSettingsComponent(props: Props): JSX.Element {
         onSwitchSmoothImage,
         onChangeCanvasBackgroundColor,
         onSwitchShowingDeletedFrames,
+        onSwitchShowOriginalImage,
     } = props;
 
     const minFrameStep = 2;
@@ -217,6 +221,26 @@ export default function PlayerSettingsComponent(props: Props): JSX.Element {
                     </Row>
                     <Row>
                         <Text type='secondary'>You will be able to navigate and restore deleted frames</Text>
+                    </Row>
+                </Col>
+            </Row>
+            <Row justify='start'>
+                <Col span={7}>
+                    <Row className='cvat-player-settings-show-original'>
+                        <Col span={24} className='cvat-player-settings-show-original-checkbox'>
+                            <Checkbox
+                                className='cvat-text-color'
+                                checked={showOriginalImage}
+                                onChange={(event: CheckboxChangeEvent): void => {
+                                    onSwitchShowOriginalImage(event.target.checked);
+                                }}
+                            >
+                                Show original image
+                            </Checkbox>
+                        </Col>
+                        <Col span={24}>
+                            <Text type='secondary'> Downloads the original image instead of the compressed version </Text>
+                        </Col>
                     </Row>
                 </Col>
             </Row>