+++ cvat-ui/src/reducers/index.ts
@@ -770,6 +770,7 @@ export interface PlayerSettingsState {
     rotateAll: boolean;
     smoothImage: boolean;
     showDeletedFrames: boolean;
+    showOriginalImage: boolean;
     grid: boolean;
     gridSize: number;
     gridColor: GridColor;